Revert "Fix more overflows"
This reverts commit cc6135dd02ab988e59bbe1f257d249fc46fbd8de.

+++ tile.cpp
@@ -1265,8 +1265,8 @@ void *run_prefilter(void *v) {
 		// Offset from tile coordinates back to world coordinates
 		unsigned sx = 0, sy = 0;
 		if (rpa->z != 0) {
-			sx = (long long) rpa->tx << (GLOBAL_DETAIL - rpa->z);
-			sy = (long long) rpa->ty << (GLOBAL_DETAIL - rpa->z);
+			sx = rpa->tx << (GLOBAL_DETAIL - rpa->z);
+			sy = rpa->ty << (GLOBAL_DETAIL - rpa->z);
 		}
 		for (size_t i = 0; i < tmp_feature.geometry.size(); i++) {
 			tmp_feature.geometry[i].x += sx;
